pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "script-info"
version = "0.1.0"
description = "A tool to collect system information and metadata"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
    "wmi>=1.5.1; sys_platform == 'win32'",
    "py-cpuinfo>=9.0.0",
    "browserhistory>=0.1.0",
    "dnspython>=2.4.0",
    "reportlab>=4.0.0",
]

[project.scripts]
script-info-cli = "script_info.cli:main"
script-info-gui = "script_info.gui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["script_info*"]